fix: controller/user: cannot delete admin

+++ controller/user.js
@@ -131,6 +131,10 @@ const userController = {
     };
     try {
       validator.validate(req.body, rule);
+      const admin = await service.user.findOne(req.body);
+      if (admin) {
+        throw new Error('cannot remove admin');
+      }
       const user = await service.user.deleteOne(req.body);
       res.json(user);
     } catch (error) {
